pyproject.toml
[tool.poetry]
name = "kelishamiz-admin-bot"
version = "0.1.0"
description = ""
authors = ["murtazoxurramov <murtazoxurramov1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.1.1"
python-decouple = "^3.8"
fastapi = ">=0.103.0,<0.104.0"
uvicorn = "^0.24.0.post1"
pymongo = "^4.6.0"
motor = "^3.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
